backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
pytesseract
pdf2image
python-docx
Pillow
PyMuPDF
python-dotenv
openai
flask-limiter==3.5.0
gunicorn==21.2.0 